requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.335
openai==1.3.5
qdrant-client==1.7.0
python-dotenv==1.0.0
pydantic==2.5.0
tiktoken==0.5.2
unstructured>=0.12.0
pypdf==4.0.0
pytest==7.4.3
httpx==0.25.2
google-generativeai==0.8.4
huggingface_hub>=0.34.0
sentence-transformers==5.2.0
transformers==4.57.3
torch>=1.11.0